@@ -1,66 +1,67 @@
-"""reset initial migration
+"""
+reset initial migration
 
 Revision ID: 37a5f869f63c
 Revises: 
 Create Date: 2024-12-27 23:10:29.728781
 
 """
-from typing import Sequence, Union
+from collections.abc import Sequence
 
-from alembic import op
 import sqlalchemy as sa
 
+from alembic import op
 
 # revision identifiers, used by Alembic.
-revision: str = '37a5f869f63c'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
+revision: str = "37a5f869f63c"
+down_revision: str | None = None
+branch_labels: str | Sequence[str] | None = None
+depends_on: str | Sequence[str] | None = None
 
 
 def upgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('config',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('key', sa.String(), nullable=False),
-    sa.Column('value', sa.String(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
+    op.create_table("config",
+    sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column("key", sa.String(), nullable=False),
+    sa.Column("value", sa.String(), nullable=True),
+    sa.PrimaryKeyConstraint("id")
     )
-    with op.batch_alter_table('config', schema=None) as batch_op:
-        batch_op.create_index(batch_op.f('ix_config_id'), ['id'], unique=False)
-        batch_op.create_index(batch_op.f('ix_config_key'), ['key'], unique=False)
+    with op.batch_alter_table("config", schema=None) as batch_op:
+        batch_op.create_index(batch_op.f("ix_config_id"), ["id"], unique=False)
+        batch_op.create_index(batch_op.f("ix_config_key"), ["key"], unique=False)
 
-    op.create_table('images',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('filename', sa.String(), nullable=False),
-    sa.Column('filepath', sa.String(), nullable=False),
-    sa.Column('filetype', sa.String(), nullable=False),
-    sa.Column('thumbnail_path', sa.String(), nullable=True),
-    sa.Column('width', sa.Integer(), nullable=True),
-    sa.Column('height', sa.Integer(), nullable=True),
-    sa.Column('aspect_width', sa.Integer(), nullable=True),
-    sa.Column('aspect_height', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
+    op.create_table("images",
+    sa.Column("id", sa.Integer(), autoincrement=True, nullable=False),
+    sa.Column("filename", sa.String(), nullable=False),
+    sa.Column("filepath", sa.String(), nullable=False),
+    sa.Column("filetype", sa.String(), nullable=False),
+    sa.Column("thumbnail_path", sa.String(), nullable=True),
+    sa.Column("width", sa.Integer(), nullable=True),
+    sa.Column("height", sa.Integer(), nullable=True),
+    sa.Column("aspect_width", sa.Integer(), nullable=True),
+    sa.Column("aspect_height", sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint("id")
     )
-    with op.batch_alter_table('images', schema=None) as batch_op:
-        batch_op.create_index('ix_images_filename', ['filename'], unique=False)
-        batch_op.create_index('ix_images_filepath', ['filepath'], unique=False)
-        batch_op.create_index(batch_op.f('ix_images_id'), ['id'], unique=False)
+    with op.batch_alter_table("images", schema=None) as batch_op:
+        batch_op.create_index("ix_images_filename", ["filename"], unique=False)
+        batch_op.create_index("ix_images_filepath", ["filepath"], unique=False)
+        batch_op.create_index(batch_op.f("ix_images_id"), ["id"], unique=False)
 
     # ### end Alembic commands ###
 
 
 def downgrade() -> None:
     # ### commands auto generated by Alembic - please adjust! ###
-    with op.batch_alter_table('images', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_images_id'))
-        batch_op.drop_index('ix_images_filepath')
-        batch_op.drop_index('ix_images_filename')
+    with op.batch_alter_table("images", schema=None) as batch_op:
+        batch_op.drop_index(batch_op.f("ix_images_id"))
+        batch_op.drop_index("ix_images_filepath")
+        batch_op.drop_index("ix_images_filename")
 
-    op.drop_table('images')
-    with op.batch_alter_table('config', schema=None) as batch_op:
-        batch_op.drop_index(batch_op.f('ix_config_key'))
-        batch_op.drop_index(batch_op.f('ix_config_id'))
+    op.drop_table("images")
+    with op.batch_alter_table("config", schema=None) as batch_op:
+        batch_op.drop_index(batch_op.f("ix_config_key"))
+        batch_op.drop_index(batch_op.f("ix_config_id"))
 
-    op.drop_table('config')
+    op.drop_table("config")
     # ### end Alembic commands ###
