@@ -1,7 +1,11 @@
+import logging
+
 from pydantic_settings import BaseSettings, SettingsConfigDict
 
 
 class Settings(BaseSettings):
+    """Settings for the Frame Gallery application."""
+
     model_config = SettingsConfigDict(env_file=".env", env_file_encoding="utf-8")
 
     app_name: str = "Frame Gallery"
@@ -11,9 +15,9 @@ class Settings(BaseSettings):
     db_url: str = "sqlite:///./data/framegallery.db"
     filesystem_refresh_interval: int = 600
     slideshow_interval: int = 180
-    log_level: str = "INFO"
+    log_level: str = "DEBUG"
 
 
 settings = Settings()
 
-print(f"Settings at boot: {settings.model_dump()}")
+logging.warning("Settings at boot: %s", settings.model_dump())
