@@ -1,50 +1,59 @@
 import asyncio
 import logging
 import os
+from collections.abc import AsyncGenerator
 from contextlib import asynccontextmanager
-from typing import Dict
+from pathlib import Path
+from typing import Annotated
 
-from fastapi import Depends, FastAPI
+from fastapi import Depends, FastAPI, HTTPException
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.requests import Request
 from fastapi.staticfiles import StaticFiles
 from fastapi.templating import Jinja2Templates
 from sqlalchemy.orm import Session
 
-import framegallery.crud as crud
-import framegallery.models as models
-from framegallery.schemas import ConfigResponse, Image
+from framegallery import crud, models, schemas
+from framegallery.config import settings
 from framegallery.configuration.update_current_active_image_config_listener import (
     UpdateCurrentActiveImageConfigListener,
 )
-from framegallery.repository.image_repository import ImageRepository
-from framegallery.repository.config_repository import ConfigRepository, ConfigKey
-from framegallery.frame_connector.frame_connector import FrameConnector
-from framegallery.config import settings
 from framegallery.database import engine, get_db
+from framegallery.dependencies import get_config_repository, get_filter_repository, get_slideshow_instance
+from framegallery.frame_connector.frame_connector import FrameConnector, api_version
 from framegallery.frame_connector.status import SlideshowStatus, Status
-from framegallery.frame_connector.frame_connector import api_version
 from framegallery.importer2.importer import Importer
-from framegallery.slideshow.slideshow import Slideshow, get_slideshow
+from framegallery.logging_config import setup_logging
+from framegallery.repository.config_repository import ConfigKey, ConfigRepository
+from framegallery.repository.filter_repository import FilterRepository
+from framegallery.repository.image_repository import ImageRepository
+from framegallery.routers import config_router, filters_router
+from framegallery.schemas import ConfigResponse, Filter, Image
+from framegallery.slideshow.slideshow import Slideshow
+
+logger = setup_logging(log_level=settings.log_level)
 
 models.Base.metadata.create_all(bind=engine)
-
-logging.basicConfig(level=logging.getLevelName(settings.log_level))
 
 # Create Frame TV Connector
 frame_connector = FrameConnector(settings.tv_ip_address, settings.tv_port)
 
+background_tasks = set()
 
 # Background task to run the filesystem sync
-async def run_importer_periodically(db: Session):
+async def run_importer_periodically(db: Session) -> None:
+    """Run the importer periodically to synchronize the filesystem with the database."""
+    logging.info("Inside run_importer_periodically")
+
     importer = Importer(settings.gallery_path, db)
     while True:
-        logging.debug("Running importer")
+        logging.info("Running importer now")
         await importer.synchronize_files()
         await asyncio.sleep(settings.filesystem_refresh_interval)
 
 
-async def update_slideshow_periodically(slideshow: Slideshow):
+async def update_slideshow_periodically(slideshow: Slideshow) -> None:
+    """Update the slideshow periodically."""
     while True:
         logging.debug("Updating slideshow")
         await slideshow.update_slideshow()
@@ -52,25 +61,36 @@ async def update_slideshow_periodically(slideshow: Slideshow):
 
 
 @asynccontextmanager
-async def lifespan(app: FastAPI):
+async def lifespan(app: FastAPI) -> AsyncGenerator[None, any]:
+    """Run tasks on startup and shutdown."""
+    logger.info("logger - Inside lifespan")
     await frame_connector.get_active_item_details()
 
     # Create a database session and run the importer periodically
     db = next(get_db())
-    asyncio.create_task(run_importer_periodically(db))
+    logger.info("Scheduling the filesystem importer task")
+    image_importer = asyncio.create_task(run_importer_periodically(db))
+    background_tasks.add(image_importer)
+    image_importer.add_done_callback(background_tasks.discard)
 
     image_repository = ImageRepository(db)
-    slideshow = next(get_slideshow(image_repository))
-    asyncio.create_task(update_slideshow_periodically(slideshow))
+    config_repository = ConfigRepository(db)
+    filter_repository = FilterRepository(db)
+    slideshow = get_slideshow_instance(image_repository, config_repository, filter_repository)
+    logger.info("Scheduling the slideshow updater")
+    slideshow_updater = asyncio.create_task(update_slideshow_periodically(slideshow))
+    background_tasks.add(slideshow_updater)
+    slideshow_updater.add_done_callback(background_tasks.discard)
 
     config_repository = ConfigRepository(db)
-    update_current_active_image_config_listener = (
-        UpdateCurrentActiveImageConfigListener(config_repository)
-    )
+    # Store the listener in the app state so that it doesn't get garbage collected
+    app.state.update_active_image_in_config_listener = UpdateCurrentActiveImageConfigListener(config_repository)
 
     yield
 
 
+logger.info("logger - Before FastAPI launch")
+logging.info("logging - Before FastAPI launch")
 app = FastAPI(lifespan=lifespan)
 
 origins = [
@@ -87,11 +107,10 @@ app.add_middleware(
     allow_headers=["*"],
 )
 
-
 # Sets the templates directory to the `build` folder from `npm run build`
 # this is where you'll find the index.html file.
 templates = Jinja2Templates(directory="./ui/build")
-# templates = Jinja2Templates(directory="./ui/templates")
+# Uncomment this when running from npm dev: templates = Jinja2Templates(directory="./ui/templates")
 
 # Mounts the `static` folder within the `build` folder to the `/static` route.
 app.mount("/static", StaticFiles(directory="./ui/build/static"), "static")
@@ -99,8 +118,8 @@ app.mount("/images", StaticFiles(directory=settings.gallery_path), "images")
 
 
 @app.get("/api/status")
-async def status(request: Request) -> Status:
-    # Stub the response for now
+async def status() -> Status:
+    """Return the application state. Stubbed for now."""
     return Status(
         tv_on=True,
         art_mode_supported=True,
@@ -109,8 +128,9 @@ async def status(request: Request) -> Status:
     )
 
 
-@app.get("/api/available-images")
-async def available_images(request: Request, db: Session = Depends(get_db)):
+@app.get("/api/available-images", response_model=list[schemas.Image])
+async def available_images(db: Annotated[Session,  Depends(get_db)]) -> list[models.Image]:
+    """Get a list of all available images."""
     images = crud.get_images(db)
 
     for image in images:
@@ -124,17 +144,17 @@ async def available_images(request: Request, db: Session = Depends(get_db)):
 """
 Retrieves a list of (nested) folders in the gallery path, that can be used for filtering in the UI.
 """
-
-
 @app.get("/api/albums")
-async def get_albums(request: Request):
-    def build_tree(path: str) -> Dict:
+async def get_albums() -> dict:
+    """Get a directory tree of gallery albums."""
+    def build_tree(path: str) -> dict:
+        """Build a directory tree of gallery albums."""
         tree = {"id": "/", "name": "/", "label": "/", "children": []}
         for root, dirs, _ in os.walk(path):
             folder = root.replace(path, "").strip(os.sep)
             subtree = tree["children"]
             if folder:
-                for part in folder.split(os.sep):
+                for part in Path(folder).parts:
                     found = next((item for item in subtree if item["id"] == part), None)
                     if not found:
                         found = {
@@ -159,88 +179,115 @@ async def get_albums(request: Request):
 
     return build_tree(settings.gallery_path)
 
+@app.post("/api/active-art/{image_id}")
+async def select_art(image_id: int, db: Annotated[Session, Depends(get_db)],
+                     slideshow: Annotated[Slideshow, Depends(get_slideshow_instance)]) -> Image:
+    """Set the active item."""
+    image = crud.get_image_by_id(db, image_id)
+    if not image:
+        raise HTTPException(status_code=404, detail=f"Image with ID {image_id} not found")
 
-# """
-# Sets the active item
-# """
-# @app.post("/api/active-art/{id}")
-# async def select_art(request: Request, id: int, db: Session = Depends(get_db), slideshow: Slideshow = Depends(get_slideshow)):
-#     image = crud.get_image_by_id(db, id)
-#     if not image:
-#         raise HTTPException(status_code=404, detail=f"Image with ID {id} not found")
-#
-#     await slideshow.set_slideshow_active_image(image)
-#
-#     # Update current_active_art
-#     await active_art()
-#
-#     return image
+    await slideshow.set_slideshow_active_image(image)
+
+    return image
 
 
 @app.get("/api/slideshow")
-async def get_slideshow_status(
-    request: Request, db: Session = Depends(get_db)
-) -> SlideshowStatus:
+async def get_slideshow_status(db: Annotated[Session,  Depends(get_db)]) -> SlideshowStatus:
+    """Get the current slideshow status."""
     config_repo = ConfigRepository(db)
-    slideshow_status = config_repo.get_or(ConfigKey.SLIDESHOW_ENABLED, True)
+    slideshow_status = config_repo.get_or(ConfigKey.SLIDESHOW_ENABLED, default_value=True)
     return SlideshowStatus(
         enabled=slideshow_status.value == "true", interval=settings.slideshow_interval
     )
 
 
 @app.post("/api/slideshow/enable")
-async def enable_slideshow(request: Request, db: Session = Depends(get_db)):
+async def enable_slideshow(db: Annotated[Session,  Depends(get_db)]) -> dict:
+    """Enable the slideshow."""
     config_repo = ConfigRepository(db)
-    config_repo.set(ConfigKey.SLIDESHOW_ENABLED, True)
+    config_repo.set(ConfigKey.SLIDESHOW_ENABLED, value=True)
 
     return {}
 
 
 @app.post("/api/slideshow/disable")
-async def enable_slideshow(request: Request, db: Session = Depends(get_db)):
+async def disable_slideshow(db: Annotated[Session,  Depends(get_db)]) -> dict:
+    """Disable the slideshow."""
     config_repo = ConfigRepository(db)
-    config_repo.set(ConfigKey.SLIDESHOW_ENABLED, False)
+    config_repo.set(ConfigKey.SLIDESHOW_ENABLED, value=False)
 
     return {}
 
 
 @app.get("/api/settings")
 async def get_settings(
-    request: Request, db: Session = Depends(get_db)
-) -> ConfigResponse:
+    db: Annotated[Session, Depends(get_db)],
+    filter_repository: Annotated[FilterRepository, Depends(get_filter_repository)]
+    ) -> ConfigResponse:
+    """Get the current settings."""
     config_repo = ConfigRepository(db)
-    active_image_id = config_repo.get_or(ConfigKey.CURRENT_ACTIVE_IMAGE, None).value
+    active_image_id = config_repo.get_or(ConfigKey.CURRENT_ACTIVE_IMAGE, default_value=None).value
     active_image = crud.get_image_by_id(db, int(active_image_id))
+    if active_image:
+        active_image = Image.model_validate(active_image)
+
+    active_filter = None
+    active_filter_id = config_repo.get_or(
+            ConfigKey.ACTIVE_FILTER, default_value=None
+        ).value
+    if active_filter_id is not None:
+        active_filter = filter_repository.get_filter(int(active_filter_id))
+    if active_filter:
+        active_filter = Filter.model_validate(active_filter)
+
     config = {
-        "slideshow_enabled": config_repo.get_or(
-            ConfigKey.SLIDESHOW_ENABLED, True
-        ).value,
+        "slideshow_enabled": config_repo.get_or(ConfigKey.SLIDESHOW_ENABLED, default_value=True).value,
         "slideshow_interval": settings.slideshow_interval,
-        "current_active_image": Image.model_validate(active_image),
-        "current_active_image_since": config_repo.get_or(
-            ConfigKey.CURRENT_ACTIVE_IMAGE_SINCE, None
-        ).value,
+        "current_active_image": active_image,
+        "current_active_image_since":
+            config_repo.get_or(ConfigKey.CURRENT_ACTIVE_IMAGE_SINCE, default_value=None).value,
+        "active_filter": active_filter
     }
 
     return ConfigResponse(**config)
 
 
+@app.post("/api/images/next")
+async def next_image(slideshow: Annotated[Slideshow, Depends(get_slideshow_instance)]) -> Image:
+    """Advance to the next image in the slideshow."""
+    return await slideshow.update_slideshow()
+
+
+# Include routers for modular API endpoints
+app.include_router(filters_router)
+app.include_router(config_router)
+
 # Defines a route handler for `/*` essentially.
 # NOTE: this needs to be the last route defined b/c it's a catch all route
-@app.get("/{rest_of_path:path}")
-async def react_app(req: Request, rest_of_path: str, db: Session = Depends(get_db)):
-    config_repo = ConfigRepository(db)
+@app.get("/{rest_of_path:path}", response_model=None)
+async def react_app(req: Request,
+                    config_repo: Annotated[ConfigRepository, Depends(get_config_repository)],
+                    filter_repository: Annotated[FilterRepository, Depends(get_filter_repository)]
+                    ) -> templates.TemplateResponse:
+    """Render the React app."""
+    active_filter_id = config_repo.get_or(
+            ConfigKey.ACTIVE_FILTER, default_value=None
+        ).value
+    active_filter = filter_repository.get_filter(int(active_filter_id))
+
     config = {
         "slideshow_enabled": config_repo.get_or(
-            ConfigKey.SLIDESHOW_ENABLED, True
+            ConfigKey.SLIDESHOW_ENABLED, default_value=True
         ).value,
         "slideshow_interval": settings.slideshow_interval,
         "current_active_image": config_repo.get_or(
-            ConfigKey.CURRENT_ACTIVE_IMAGE, None
+            ConfigKey.CURRENT_ACTIVE_IMAGE, default_value=None
         ).value,
         "current_active_image_since": config_repo.get_or(
-            ConfigKey.CURRENT_ACTIVE_IMAGE_SINCE, None
+            ConfigKey.CURRENT_ACTIVE_IMAGE_SINCE, default_value=None
         ).value,
+        "active_filter": active_filter
     }
 
     return templates.TemplateResponse("index.html", {"request": req, "config": config})
