@@ -1,13 +1,14 @@
 from sqlalchemy import Index, Integer, String
-from sqlalchemy.orm import DeclarativeBase, Mapped
-from sqlalchemy.orm import mapped_column
+from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
 class Base(DeclarativeBase):
-    pass
+    """Base class for all models."""
 
 
 class Image(Base):
+    """Class for images."""
+
     __tablename__ = "images"
     __table_args__ = (
         Index("ix_images_filename", "filename"),
@@ -25,7 +26,18 @@ class Image(Base):
 
 
 class Config(Base):
+    """Configuration settings."""
+
     __tablename__ = "config"
     id: Mapped[int] = mapped_column(primary_key=True, index=True, autoincrement=True)
     key: Mapped[str] = mapped_column(String, index=True)
     value: Mapped[str] = mapped_column(String, nullable=True)
+
+
+class Filter(Base):
+    """Class for image filters."""
+
+    __tablename__ = "filters"
+    id: Mapped[int] = mapped_column(primary_key=True, index=True, autoincrement=True)
+    name: Mapped[str] = mapped_column(String, index=True)
+    query: Mapped[str] = mapped_column(String, nullable=True)
