@@ -3,12 +3,14 @@ import { API_BASE_URL } from './App';
 import Image from './models/Image';
 
 // Define the shape of your settings
+import { Filter } from './components/Filters/Filter';
+
 interface Settings {
-  // [key: string]: any;  // Replace with your specific settings structure
-  slideshow_enabled: Boolean;
-  slideshow_interval: Number;
+  slideshow_enabled: boolean;
+  slideshow_interval: number;
   current_active_image: Image;
   current_active_image_since: string | null;
+  active_filter: Filter | null;
 }
 
 // Define the context value type
@@ -16,7 +18,7 @@ interface SettingsContextValue {
   settings: Settings | null;
   loading: boolean;
   error: string | null;
-  updateSetting: (key: string, value: any) => Promise<boolean>; // eslint-disable-line no-unused-vars
+  updateSetting: (key: string, value: any) => Promise<boolean>; // eslint-disable-line no-unused-vars, @typescript-eslint/no-explicit-any
 }
 
 // Create the context with a default value
@@ -64,6 +66,7 @@ export const SettingsProvider: React.FC<SettingsProviderProps> = ({ children }) 
   }, []);
 
   // Method to update a specific setting
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   const updateSetting = async (key: string, value: any): Promise<boolean> => {
     try {
       const response = await fetch('/api/settings', {
